+++ maze/src/cast_ray.c
@@ -2,18 +2,18 @@
 #include <stdlib.h>
 
 /* Draw the maze */
-int *cast_ray(float ray_angle, Player_POV *player, char (*map)[MAP_WIDTH])
+float *cast_ray(float ray_angle, Player_POV *player, char (*map)[MAP_WIDTH])
 {
-    static int coords[2];
-    int raw_distance;
-    int skewed_distance;
-    int col_length;
+    static float coords[2];
+    float raw_distance;
+    float skewed_distance;
+    float col_length;
 
     /* Debugging: */
     if (HIDE_INFO) {
-        printf("Player x: %i\n", player->x_coord);
-        printf("Player y: %i\n", player->y_coord);
-        printf("Player angle: %i\n", player->angle);
+        printf("Player x: %f\n", player->x_coord);
+        printf("Player y: %f\n", player->y_coord);
+        printf("Player angle: %f\n", player->angle);
         printf("char: %c\n", map[0][0]);
         printf("ray angle: %f\n", ray_angle);
     } else if (DEBUG) {
@@ -28,9 +28,9 @@ int *cast_ray(float ray_angle, Player_POV *player, char (*map)[MAP_WIDTH])
         return coords;
     }
     skewed_distance = player->angle - ray_angle;
-    printf("distance: %i, skewed_distance: %i\n", raw_distance, skewed_distance);
+    printf("distance: %f, skewed_distance: %f\n", raw_distance, skewed_distance);
     col_length = skewed_distance * SCALE_VALUE;
-    printf("wall slice (column) length: %i\n", col_length);
+    printf("wall slice (column) length: %f\n", col_length);
     coords[0] = WINDOW_HEIGHT / 2 - col_length / 2;
     coords[1] = coords[0] + col_length;
